+++ fractal_server/app/runner/v2/__init__.py
@@ -1,4 +1,3 @@
-from .images import *  # noqa: F403, F401
 from .models import *  # noqa: F403, F401
 from .runner import *  # noqa: F403, F401
 from .runner_functions import *  # noqa: F403, F401
