+++ fractal_server/app/runner/v2/runner.py
@@ -1,10 +1,10 @@
 from copy import copy
 from copy import deepcopy
 
-from .images import _deduplicate_list
-from .images import filter_images
-from .images import find_image_by_path
-from .images import SingleImage
+from ....images import deduplicate_list
+from ....images import filter_images
+from ....images import find_image_by_path
+from ....images import SingleImage
 from .models import Dataset
 from .models import DictStrAny
 from .models import WorkflowTask
@@ -119,7 +119,7 @@ def execute_tasks_v2(
                             **wftask.args,
                         )
                     )
-                list_function_kwargs = _deduplicate_list(list_function_kwargs)
+                list_function_kwargs = deduplicate_list(list_function_kwargs)
 
                 current_task_images = [
                     find_image_by_path(
@@ -128,7 +128,7 @@ def execute_tasks_v2(
                     for kwargs in list_function_kwargs
                 ]
 
-                current_task_images = _deduplicate_list(current_task_images)
+                current_task_images = deduplicate_list(current_task_images)
 
             task_output = _run_parallel_task(
                 task=task,
@@ -160,7 +160,7 @@ def execute_tasks_v2(
                 image=image, new_attributes=new_filters
             )
             added_images[ind] = updated_image
-        added_images = _deduplicate_list(added_images)
+        added_images = deduplicate_list(added_images)
 
         # Get removed images
         removed_images = task_output.removed_images or []
