--- fractal_server/app/schemas/v2/_validators_v2.py
@@ -1,17 +1,0 @@
-from typing import Any
-from typing import Union
-
-
-def val_scalar_dict(attribute: str):
-    def val(
-        dict_str_any: dict[str, Any],
-    ) -> dict[str, Union[int, float, str, bool, None]]:
-        for key, value in dict_str_any.items():
-            if not isinstance(value, (int, float, str, bool, type(None))):
-                raise ValueError(
-                    f"{attribute}[{key}] must be a scalar (int, float, str, "
-                    f"bool, or None). Given {value} ({type(value)})"
-                )
-        return dict_str_any
-
-    return val
